requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
nltk
tensorflow
tweepy
scipy
python-dotenv
